requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
weasyprint
jinja2
python-multipart
resend
slowapi
pytest
pytest-asyncio
httpx
